+++ idea/src/org/jetbrains/kotlin/idea/refactoring/memberInfo/KotlinMemberInfo.kt
@@ -16,24 +16,32 @@
 
 package org.jetbrains.kotlin.idea.refactoring.memberInfo
 
+import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiField
 import com.intellij.psi.PsiMember
+import com.intellij.psi.PsiMethod
 import com.intellij.refactoring.RefactoringBundle
 import com.intellij.refactoring.classMembers.MemberInfoBase
 import com.intellij.refactoring.util.classMembers.MemberInfo
-import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.getRepresentativeLightMethod
 import org.jetbrains.kotlin.asJava.toLightClass
+import org.jetbrains.kotlin.asJava.toLightElements
 import org.jetbrains.kotlin.asJava.unwrapped
 import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.descriptors.MemberDescriptor
 import org.jetbrains.kotlin.descriptors.Modality
-import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptor
+import org.jetbrains.kotlin.idea.refactoring.isInterfaceClass
 import org.jetbrains.kotlin.idea.util.IdeDescriptorRenderers
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.renderer.DescriptorRendererModifier
+import org.jetbrains.kotlin.utils.addToStdlib.firstIsInstanceOrNull
 import org.jetbrains.kotlin.utils.addToStdlib.singletonOrEmptySet
 
-class KotlinMemberInfo(member: KtNamedDeclaration, val isSuperClass: Boolean = false) : MemberInfoBase<KtNamedDeclaration>(member) {
+class KotlinMemberInfo @JvmOverloads constructor(
+        member: KtNamedDeclaration,
+        val isSuperClass: Boolean = false,
+        val isCompanionMember: Boolean = false
+) : MemberInfoBase<KtNamedDeclaration>(member) {
     companion object {
         private val RENDERER = IdeDescriptorRenderers.SOURCE_CODE_SHORT_NAMES_IN_TYPES.withOptions {
             modifiers = DescriptorRendererModifier.INNER.singletonOrEmptySet()
@@ -41,11 +49,11 @@ class KotlinMemberInfo(member: KtNamedDeclaration, val isSuperClass: Boolean = f
     }
 
     init {
-        val memberDescriptor = member.resolveToDescriptor()
+        val memberDescriptor = member.resolveToDescriptorWrapperAware()
         isStatic = member.parent is KtFile
 
-        if (member is KtClass && isSuperClass) {
-            if (member.isInterface()) {
+        if ((member is KtClass || member is KtPsiClassWrapper) && isSuperClass) {
+            if (member.isInterfaceClass()) {
                 displayName = RefactoringBundle.message("member.info.implements.0", member.name)
                 overrides = false
             }
@@ -59,6 +67,9 @@ class KotlinMemberInfo(member: KtNamedDeclaration, val isSuperClass: Boolean = f
             if (memberDescriptor is MemberDescriptor && memberDescriptor.modality == Modality.ABSTRACT) {
                 displayName = "abstract $displayName"
             }
+            if (isCompanionMember) {
+                displayName = "companion $displayName"
+            }
 
             val overriddenDescriptors = (memberDescriptor as? CallableMemberDescriptor)?.overriddenDescriptors ?: emptySet()
             if (overriddenDescriptors.isNotEmpty()) {
@@ -68,16 +79,24 @@ class KotlinMemberInfo(member: KtNamedDeclaration, val isSuperClass: Boolean = f
     }
 }
 
-fun KotlinMemberInfo.toJavaMemberInfo(): MemberInfo? {
-    val declaration = member
-    val psiMember: PsiMember? = when (declaration) {
+fun lightElementForMemberInfo(declaration: KtNamedDeclaration?): PsiMember? {
+    return when (declaration) {
         is KtNamedFunction -> declaration.getRepresentativeLightMethod()
-        is KtProperty -> declaration.getRepresentativeLightMethod() ?: LightClassUtil.getLightClassPropertyMethods(declaration).backingField
+        is KtProperty, is KtParameter -> declaration.toLightElements().let {
+            it.firstIsInstanceOrNull<PsiMethod>() ?: it.firstIsInstanceOrNull<PsiField>()
+        } as PsiMember?
         is KtClassOrObject -> declaration.toLightClass()
+        is KtPsiClassWrapper -> declaration.psiClass
         else -> null
     }
-    val info = MemberInfo(psiMember ?: return null, isSuperClass, null)
+}
+
+fun MemberInfoBase<out KtNamedDeclaration>.toJavaMemberInfo(): MemberInfo? {
+    val declaration = member
+    val psiMember: PsiMember? = lightElementForMemberInfo(declaration)
+    val info = MemberInfo(psiMember ?: return null, psiMember is PsiClass && overrides != null, null)
     info.isToAbstract = isToAbstract
+    info.isChecked = isChecked
     return info
 }