+++ idea/src/org/jetbrains/kotlin/idea/refactoring/pullUp/KotlinPullUpDialog.kt
@@ -22,12 +22,12 @@ import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiComment
 import com.intellij.psi.PsiNamedElement
 import com.intellij.refactoring.JavaRefactoringSettings
-import com.intellij.refactoring.classMembers.AbstractMemberInfoModel
 import com.intellij.refactoring.classMembers.MemberInfoChange
 import com.intellij.refactoring.classMembers.MemberInfoModel
 import com.intellij.refactoring.memberPullUp.PullUpProcessor
 import com.intellij.refactoring.util.DocCommentPolicy
 import org.jetbrains.kotlin.asJava.toLightClass
+import org.jetbrains.kotlin.idea.refactoring.isCompanionMemberOf
 import org.jetbrains.kotlin.idea.refactoring.isInterfaceClass
 import org.jetbrains.kotlin.idea.refactoring.memberInfo.*
 import org.jetbrains.kotlin.psi.*
@@ -44,7 +44,16 @@ class KotlinPullUpDialog(
         init()
     }
 
-    private inner class MemberInfoModelImpl : AbstractMemberInfoModel<KtNamedDeclaration, KotlinMemberInfo>() {
+    private inner class MemberInfoModelImpl(
+            originalClass: KtClassOrObject,
+            superClass: PsiNamedElement?,
+            interfaceContainmentVerifier: (KtNamedDeclaration) -> Boolean
+    ) : KotlinUsesAndInterfacesDependencyMemberInfoModel<KtNamedDeclaration, KotlinMemberInfo>(
+            originalClass,
+            superClass,
+            false,
+            interfaceContainmentVerifier
+    ) {
         private var lastSuperClass: PsiNamedElement? = null
 
         // Abstract members remain abstract
@@ -62,33 +71,43 @@ class KotlinPullUpDialog(
             val superClass = superClass ?: return false
             if (superClass is PsiClass) return false
             if (superClass !is KtClass) return false
-            if (!superClass.isInterface()) return true
 
             val member = memberInfo.member
-            return member is KtNamedFunction || (member is KtProperty && !member.mustBeAbstractInInterface())
+            if (member.isCompanionMemberOf(sourceClass)) return false
+
+            if (!superClass.isInterface()) return true
+
+            return member is KtNamedFunction || (member is KtProperty && !member.mustBeAbstractInInterface()) || member is KtParameter
         }
 
         override fun isAbstractWhenDisabled(memberInfo: KotlinMemberInfo): Boolean {
             val member = memberInfo.member
-            return (member is KtProperty && superClass !is PsiClass) || (member is KtNamedFunction && superClass is PsiClass)
+            if (member.isCompanionMemberOf(sourceClass)) return false
+            return ((member is KtProperty || member is KtParameter) && superClass !is PsiClass)
+                   || (member is KtNamedFunction && superClass is PsiClass)
         }
 
         override fun isMemberEnabled(memberInfo: KotlinMemberInfo): Boolean {
             val superClass = superClass ?: return false
             val member = memberInfo.member
 
-            if (superClass is PsiClass && !member.canMoveMemberToJavaClass(superClass)) return false
+            if (superClass is PsiClass) {
+                if (!member.canMoveMemberToJavaClass(superClass)) return false
+                if (member.isCompanionMemberOf(sourceClass)) return false
+            }
             if (memberInfo in memberInfoStorage.getDuplicatedMemberInfos(superClass)) return false
             if (member in memberInfoStorage.getExtending(superClass)) return false
             return true
         }
 
         override fun memberInfoChanged(event: MemberInfoChange<KtNamedDeclaration, KotlinMemberInfo>) {
+            super.memberInfoChanged(event)
             val superClass = superClass ?: return
             if (superClass != lastSuperClass) {
                 lastSuperClass = superClass
                 val isInterface = superClass is KtClass && superClass.isInterface()
                 event.changedMembers.forEach { it.isToAbstract = isInterface }
+                setSuperClass(superClass)
             }
         }
     }
@@ -102,7 +121,7 @@ class KotlinPullUpDialog(
     override fun getSuperClass() = super.getSuperClass()
 
     override fun createMemberInfoModel(): MemberInfoModel<KtNamedDeclaration, KotlinMemberInfo> =
-            MemberInfoModelImpl()
+            MemberInfoModelImpl(sourceClass, preselection, getInterfaceContainmentVerifier { selectedMemberInfos })
 
     override fun getPreselection() = mySuperClasses.firstOrNull { !it.isInterfaceClass() } ?: mySuperClasses.firstOrNull()