+++ idea/src/org/jetbrains/kotlin/idea/refactoring/pullUp/KotlinPullUpHelper.kt
@@ -34,19 +34,22 @@ import org.jetbrains.kotlin.descriptors.*
 import org.jetbrains.kotlin.idea.KotlinLanguage
 import org.jetbrains.kotlin.idea.codeInsight.shorten.addToShorteningWaitSet
 import org.jetbrains.kotlin.idea.core.dropDefaultValue
+import org.jetbrains.kotlin.idea.core.getOrCreateCompanionObject
+import org.jetbrains.kotlin.idea.core.replaced
 import org.jetbrains.kotlin.idea.intentions.setType
 import org.jetbrains.kotlin.idea.refactoring.createJavaField
 import org.jetbrains.kotlin.idea.refactoring.dropOverrideKeywordIfNecessary
+import org.jetbrains.kotlin.idea.refactoring.isCompanionMemberOf
+import org.jetbrains.kotlin.idea.refactoring.memberInfo.KtPsiClassWrapper
+import org.jetbrains.kotlin.idea.refactoring.memberInfo.toKtDeclarationWrapperAware
 import org.jetbrains.kotlin.idea.refactoring.safeDelete.removeOverrideModifier
 import org.jetbrains.kotlin.idea.util.anonymousObjectSuperTypeOrNull
 import org.jetbrains.kotlin.idea.util.psi.patternMatching.KotlinPsiUnifier
 import org.jetbrains.kotlin.lexer.KtModifierKeywordToken
 import org.jetbrains.kotlin.lexer.KtTokens
+import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.allChildren
-import org.jetbrains.kotlin.psi.psiUtil.asAssignment
-import org.jetbrains.kotlin.psi.psiUtil.getStrictParentOfType
-import org.jetbrains.kotlin.psi.psiUtil.parentsWithSelf
+import org.jetbrains.kotlin.psi.psiUtil.*
 import org.jetbrains.kotlin.resolve.BindingContext
 import org.jetbrains.kotlin.resolve.DescriptorUtils
 import org.jetbrains.kotlin.resolve.calls.callUtil.getResolvedCall
@@ -324,7 +327,8 @@ class KotlinPullUpHelper(
         return clashingSuperDescriptor.source.getPsi() as? KtCallableDeclaration
     }
 
-    private fun moveSuperInterface(member: KtClass, substitutor: PsiSubstitutor) {
+    private fun moveSuperInterface(member: PsiNamedElement, substitutor: PsiSubstitutor) {
+        val realMemberPsi = if (member is KtPsiClassWrapper) member.psiClass else member
         val classDescriptor = data.memberDescriptors[member] as? ClassDescriptor ?: return
         val currentSpecifier = data.sourceClass.getSuperTypeEntryByDescriptor(classDescriptor, data.sourceClassContext) ?: return
         when (data.targetClass) {
@@ -339,7 +343,7 @@ class KotlinPullUpHelper(
                 val sourcePsiClass = data.sourceClass.toLightClass() ?: return
                 val superRef = sourcePsiClass.implementsList
                                        ?.referenceElements
-                                       ?.firstOrNull { it.resolve()?.unwrapped == member }
+                                       ?.firstOrNull { it.resolve()?.unwrapped == realMemberPsi }
                                 ?: return
                 val superTypeForTarget = substitutor.substitute(elementFactory.createType(superRef))
 
@@ -380,11 +384,19 @@ class KotlinPullUpHelper(
         val lightMethod = member.getRepresentativeLightMethod()!!
 
         val movedMember: PsiMember = when (member) {
-            is KtProperty -> {
+            is KtProperty, is KtParameter -> {
                 val newType = substitutor.substitute(lightMethod.returnType)
                 val newField = createJavaField(member, data.targetClass)
                 newField.typeElement?.replace(elementFactory.createTypeElement(newType))
-                member.delete()
+                if (member.isCompanionMemberOf(data.sourceClass)) {
+                    newField.modifierList?.setModifierProperty(PsiModifier.STATIC, true)
+                }
+                if (member is KtParameter) {
+                    (member.parent as? KtParameterList)?.removeParameter(member)
+                }
+                else {
+                    member.delete()
+                }
                 newField
             }
             is KtNamedFunction -> {
@@ -417,9 +429,9 @@ class KotlinPullUpHelper(
     }
 
     override fun move(info: MemberInfoBase<PsiMember>, substitutor: PsiSubstitutor) {
-        val member = info.member.namedUnwrappedElement as? KtNamedDeclaration ?: return
+        val member = info.member.toKtDeclarationWrapperAware() ?: return
 
-        if (member is KtClass && info.overrides != null)  {
+        if ((member is KtClass || member is KtPsiClassWrapper) && info.overrides != null)  {
             moveSuperInterface(member, substitutor)
             return
         }
@@ -448,6 +460,8 @@ class KotlinPullUpHelper(
             val movedMember: KtCallableDeclaration
             val clashingSuper = fixOverrideAndGetClashingSuper(member, memberCopy)
 
+            val psiFactory = KtPsiFactory(member)
+
             val originalIsAbstract = member.hasModifier(KtTokens.ABSTRACT_KEYWORD)
             val toAbstract = when {
                 info.isToAbstract -> true
@@ -455,12 +469,15 @@ class KotlinPullUpHelper(
                 member is KtProperty -> member.mustBeAbstractInInterface()
                 else -> false
             }
+
+            val classToAddTo = if (member.isCompanionMemberOf(data.sourceClass)) data.targetClass.getOrCreateCompanionObject() else data.targetClass
+
             if (toAbstract) {
                 if (!originalIsAbstract) {
                     makeAbstract(memberCopy, data.memberDescriptors[member] as CallableMemberDescriptor, data.sourceToTargetClassSubstitutor, data.targetClass)
                 }
 
-                movedMember = doAddCallableMember(memberCopy, clashingSuper, data.targetClass)
+                movedMember = doAddCallableMember(memberCopy, clashingSuper, classToAddTo)
                 if (member.typeReference == null) {
                     movedMember.typeReference?.addToShorteningWaitSet()
                 }
@@ -468,8 +485,34 @@ class KotlinPullUpHelper(
                 removeOriginalMemberOrAddOverride(member)
             }
             else {
-                movedMember = doAddCallableMember(memberCopy, clashingSuper, data.targetClass)
-                member.delete()
+                movedMember = doAddCallableMember(memberCopy, clashingSuper, classToAddTo)
+                if (member is KtParameter && movedMember is KtParameter) {
+                    member.valOrVarKeyword?.delete()
+
+                    val superEntry = data.superEntryForTargetClass
+                    val superResolvedCall = data.targetClassSuperResolvedCall
+                    if (superResolvedCall != null) {
+                        val superCall = if (superEntry !is KtSuperTypeCallEntry || superEntry.valueArgumentList == null) {
+                            superEntry!!.replaced(psiFactory.createSuperTypeCallEntry("${superEntry.text}()"))
+                        } else superEntry
+                        val argumentList = superCall.valueArgumentList!!
+
+                        val parameterIndex = movedMember.parameterIndex()
+                        val prevParameterDescriptor = superResolvedCall.resultingDescriptor.valueParameters.getOrNull(parameterIndex - 1)
+                        val prevArgument = superResolvedCall.valueArguments[prevParameterDescriptor]?.arguments?.singleOrNull() as? KtValueArgument
+                        val newArgumentName = if (prevArgument != null && prevArgument.isNamed()) Name.identifier(member.name!!) else null
+                        val newArgument = psiFactory.createArgument(psiFactory.createExpression(member.name!!), newArgumentName)
+                        if (prevArgument == null) {
+                            argumentList.addArgument(newArgument)
+                        }
+                        else {
+                            argumentList.addArgumentAfter(newArgument, prevArgument)
+                        }
+                    }
+                }
+                else {
+                    member.delete()
+                }
             }
 
             if (originalIsAbstract && data.isInterfaceTarget) {