+++ idea/src/org/jetbrains/kotlin/idea/refactoring/pullUp/pullUpUtils.kt
@@ -17,14 +17,17 @@
 package org.jetbrains.kotlin.idea.refactoring.pullUp
 
 import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiMethod
+import org.jetbrains.kotlin.asJava.toLightClass
 import org.jetbrains.kotlin.descriptors.CallableMemberDescriptor
 import org.jetbrains.kotlin.descriptors.ClassDescriptor
 import org.jetbrains.kotlin.idea.codeInsight.shorten.addToShorteningWaitSet
 import org.jetbrains.kotlin.idea.core.replaced
 import org.jetbrains.kotlin.idea.intentions.setType
+import org.jetbrains.kotlin.idea.refactoring.memberInfo.KotlinMemberInfo
+import org.jetbrains.kotlin.idea.refactoring.memberInfo.lightElementForMemberInfo
 import org.jetbrains.kotlin.idea.util.IdeDescriptorRenderers
 import org.jetbrains.kotlin.idea.util.anonymousObjectSuperTypeOrNull
-import org.jetbrains.kotlin.lexer.KtModifierKeywordToken
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.resolve.BindingContext
@@ -39,10 +42,10 @@ fun KtProperty.mustBeAbstractInInterface() =
 
 fun KtNamedDeclaration.canMoveMemberToJavaClass(targetClass: PsiClass): Boolean {
     return when (this) {
-        is KtProperty -> {
+        is KtProperty, is KtParameter -> {
             if (targetClass.isInterface) return false
             if (hasModifier(KtTokens.OPEN_KEYWORD) || hasModifier(KtTokens.ABSTRACT_KEYWORD)) return false
-            if (accessors.isNotEmpty() || delegateExpression != null) return false
+            if (this is KtProperty && (accessors.isNotEmpty() || delegateExpression != null)) return false
             true
         }
         is KtNamedFunction -> valueParameters.all { it.defaultValue == null }
@@ -51,22 +54,38 @@ fun KtNamedDeclaration.canMoveMemberToJavaClass(targetClass: PsiClass): Boolean 
 }
 
 fun addMemberToTarget(targetMember: KtNamedDeclaration, targetClass: KtClassOrObject): KtNamedDeclaration {
+    if (targetMember is KtParameter) {
+        val parameterList = (targetClass as KtClass).createPrimaryConstructorIfAbsent().valueParameterList!!
+        val anchor = parameterList.parameters.firstOrNull { it.isVarArg || it.hasDefaultValue() }
+        return parameterList.addParameterBefore(targetMember, anchor)
+    }
+
     val anchor = targetClass.declarations.filterIsInstance(targetMember.javaClass).lastOrNull()
     val movedMember = when {
         anchor == null && targetMember is KtProperty -> targetClass.addDeclarationBefore(targetMember, null)
         else -> targetClass.addDeclarationAfter(targetMember, anchor)
     }
-    return movedMember as KtNamedDeclaration
+    return movedMember
 }
+
+private fun KtParameter.needToBeAbstract(targetClass: KtClassOrObject): Boolean {
+    return hasModifier(KtTokens.ABSTRACT_KEYWORD) || targetClass is KtClass && targetClass.isInterface()
+}
+
+private fun KtParameter.toProperty(): KtProperty = KtPsiFactory(this).createProperty(text)
 
 fun doAddCallableMember(
         memberCopy: KtCallableDeclaration,
         clashingSuper: KtCallableDeclaration?,
-        targetClass: KtClass): KtCallableDeclaration {
+        targetClass: KtClassOrObject
+): KtCallableDeclaration {
+    val memberToAdd = if (memberCopy is KtParameter && memberCopy.needToBeAbstract(targetClass)) memberCopy.toProperty() else memberCopy
+
     if (clashingSuper != null && clashingSuper.hasModifier(KtTokens.ABSTRACT_KEYWORD)) {
-        return clashingSuper.replaced(memberCopy)
+        return clashingSuper.replaced(if (memberToAdd is KtParameter && clashingSuper is KtProperty) memberToAdd.toProperty() else memberToAdd)
     }
-    return addMemberToTarget(memberCopy, targetClass) as KtCallableDeclaration
+
+    return addMemberToTarget(memberToAdd, targetClass) as KtCallableDeclaration
 }
 
 // TODO: Formatting rules don't apply here for some reason
@@ -85,13 +104,11 @@ fun KtClass.makeAbstract() {
 fun KtClassOrObject.getSuperTypeEntryByDescriptor(
         descriptor: ClassDescriptor,
         context: BindingContext
-): KtSuperTypeEntry? {
-    return getSuperTypeListEntries()
-            .filterIsInstance<KtSuperTypeEntry>()
-            .firstOrNull {
-                val referencedType = context[BindingContext.TYPE, it.typeReference]
-                referencedType?.constructor?.declarationDescriptor == descriptor
-            }
+): KtSuperTypeListEntry? {
+    return getSuperTypeListEntries().firstOrNull {
+        val referencedType = context[BindingContext.TYPE, it.typeReference]
+        referencedType?.constructor?.declarationDescriptor == descriptor
+    }
 }
 
 fun makeAbstract(member: KtCallableDeclaration,
@@ -136,7 +153,7 @@ fun makeAbstract(member: KtCallableDeclaration,
 }
 
 fun addSuperTypeEntry(
-        delegator: KtSuperTypeEntry,
+        delegator: KtSuperTypeListEntry,
         targetClass: KtClassOrObject,
         targetClassDescriptor: ClassDescriptor,
         context: BindingContext,
@@ -153,4 +170,16 @@ fun addSuperTypeEntry(
     val renderedType = IdeDescriptorRenderers.SOURCE_CODE.renderType(typeInTargetClass)
     val newSpecifier = KtPsiFactory(targetClass).createSuperTypeEntry(renderedType)
     targetClass.addSuperTypeListEntry(newSpecifier).addToShorteningWaitSet()
+}
+
+fun getInterfaceContainmentVerifier(getMemberInfos: () -> List<KotlinMemberInfo>): (KtNamedDeclaration) -> Boolean {
+    return result@ { member ->
+        val psiMethodToCheck = lightElementForMemberInfo(member) as? PsiMethod ?: return@result false
+        getMemberInfos().any {
+            if (!it.isSuperClass || it.overrides != false) return@any false
+
+            val psiSuperInterface = (it.member as? KtClass)?.toLightClass()
+            psiSuperInterface?.findMethodBySignature(psiMethodToCheck, true) != null
+        }
+    }
 }