+++ idea/src/org/jetbrains/kotlin/idea/refactoring/pushDown/pushDownConflictsUtils.kt
@@ -24,9 +24,11 @@ import com.intellij.usageView.UsageInfo
 import com.intellij.util.containers.MultiMap
 import org.jetbrains.kotlin.asJava.unwrapped
 import org.jetbrains.kotlin.descriptors.*
+import org.jetbrains.kotlin.idea.refactoring.memberInfo.KtPsiClassWrapper
 import org.jetbrains.kotlin.idea.refactoring.pullUp.renderForConflicts
 import org.jetbrains.kotlin.idea.references.KtReference
 import org.jetbrains.kotlin.idea.references.mainReference
+import org.jetbrains.kotlin.idea.resolve.ResolutionFacade
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.getQualifiedExpressionForReceiver
@@ -50,7 +52,7 @@ fun analyzePushDownConflicts(context: KotlinPushDownContext,
     val membersToKeepAbstract = ArrayList<KtNamedDeclaration>()
     for (info in context.membersToMove) {
         val member = info.member
-        if (!info.isChecked || (member is KtClassOrObject && info.overrides != null)) continue
+        if (!info.isChecked || ((member is KtClassOrObject || member is KtPsiClassWrapper) && info.overrides != null)) continue
 
         membersToPush += member
         if ((member is KtNamedFunction || member is KtProperty)
@@ -96,7 +98,7 @@ private fun checkConflicts(
     for (member in membersToPush) {
         checkMemberClashing(conflicts, context, member, membersToKeepAbstract, substitutor, targetClass, targetClassDescriptor)
         checkSuperCalls(conflicts, context, member, membersToPush)
-        checkExternalUsages(conflicts, context, member, targetClassDescriptor)
+        checkExternalUsages(conflicts, member, targetClassDescriptor, context.resolutionFacade)
         checkVisibility(conflicts, context, member, targetClassDescriptor)
     }
 }
@@ -168,15 +170,15 @@ private fun checkSuperCalls(
     )
 }
 
-private fun checkExternalUsages(
+internal fun checkExternalUsages(
         conflicts: MultiMap<PsiElement, String>,
-        context: KotlinPushDownContext,
-        member: KtNamedDeclaration,
-        targetClassDescriptor: ClassDescriptor
+        member: PsiElement,
+        targetClassDescriptor: ClassDescriptor,
+        resolutionFacade: ResolutionFacade
 ): Unit {
     for (ref in ReferencesSearch.search(member, member.resolveScope, false)) {
         val calleeExpr = ref.element as? KtSimpleNameExpression ?: continue
-        val resolvedCall = calleeExpr.getResolvedCall(context.resolutionFacade.analyze(calleeExpr)) ?: continue
+        val resolvedCall = calleeExpr.getResolvedCall(resolutionFacade.analyze(calleeExpr)) ?: continue
         val callElement = resolvedCall.call.callElement
         val dispatchReceiver = resolvedCall.dispatchReceiver
         if (dispatchReceiver == null || dispatchReceiver is Qualifier) continue